Decouple detection logic from listener
