+++ switch_interface/detection.py
@@ -11,33 +11,7 @@ from typing import Tuple
 
 import numpy as np
 
-# Lightweight shims to keep the old API stable
-import warnings
-
-def check_device(*args, **kwargs):
-    """Deprecated wrapper of :func:`listener.check_device`."""
-    warnings.warn(
-        "check_device moved to switch_interface.listener",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    from .listener import check_device as _real
-
-    return _real(*args, **kwargs)
-
-
-def listen(*args, **kwargs):
-    """Deprecated wrapper of :func:`listener.listen`."""
-    warnings.warn(
-        "listen moved to switch_interface.listener",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    from .listener import listen as _real
-
-    return _real(*args, **kwargs)
-
-__all__ = ["EdgeState", "detect_edges", "check_device", "listen"]
+__all__ = ["EdgeState", "detect_edges"]
 
 @dataclass
 class EdgeState:
@@ -111,6 +85,14 @@ def detect_edges(
         ),
         press_index is not None,
     )
+
+
+def __getattr__(name: str):
+    if name in {"listen", "check_device"}:
+        from . import compat
+
+        return getattr(compat, name)
+    raise AttributeError(name)
